requirements.txt
numpy==1.25.2
openai
python-dotenv==1.0.0
pandas
scipy
scikit-learn
ipykernel
matplotlib
plotly
jupyter
pdfminer.six
annoy
keybert
